# openapi.yaml
openapi: 3.1.0
info:
  title: Issuer Service API
  description: Generic swiyu Issuer Service service
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /oid4vci/api/token:
    post:
      tags:
      - Issuer OID4VCI API
      summary: Submit form data
      operationId: oauthTokenEndpoint
      parameters:
      - name: DPoP
        in: header
        required: false
        schema:
          type: string
      requestBody:
        description: OAuth 2.0 access token request to be submitted
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OauthAccessTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthToken"
        "400":
          description: "Invalid request, grant or token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "405":
          description: HTTP method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthToken"
        "415":
          description: Unsupported content type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthToken"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/api/nonce:
    post:
      tags:
      - Issuer OID4VCI API
      summary: Provide a self-contained nonce in a publicly accessible endpoint.
      description: |
        Provide nonces for proof of possessions in a manner not requiring the service to save it.
        The nonce should be used only once. The nonce has a (very) limit lifetime.
        The response should not be cached.
        For more information see <a href="https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#section-7.2">OID4VCI Nonce Endpoint specification</a></br>
        Also provides a DPoP nonce. For more details towards demonstrating proof of possession refer to <a href="https://datatracker.ietf.org/doc/html/rfc9449#name-authorization-server-provid">RFC9449</a>
      operationId: createNonce
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonceResponse"
        "405":
          description: HTTP method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonceResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/api/deferred_credential:
    post:
      tags:
      - Issuer OID4VCI API
      summary: Collect credential associated with the bearer token and the transaction
        id. This endpoint is used for deferred issuance.
      description: Issues a credential for a deferred transaction. Requires a valid
        bearer token and transaction details in the request body.
      operationId: createDeferredCredential
      parameters:
      - name: Authorization
        in: header
        description: Bearer token for authentication
        required: true
        schema:
          type: string
      - name: DPoP
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeferredCredentialEndpointRequest"
          application/jwt:
            schema:
              type: string
              description: |
                An encoded JWT as described in RFC7519, with the claims as found in the unencrypted request
        required: true
      responses:
        "200":
          description: Credential issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialEndpointResponse"
        "202":
          description: Credential issuance still pending — retry with the transaction_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeferredCredentialPendingResponse"
        "400":
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DpopError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      security:
      - bearer-jwt: []
  /oid4vci/api/credential:
    post:
      tags:
      - Issuer OID4VCI API
      summary: Collect credential associated with the bearer token with the requested
        credential properties.
      description: Issues a credential for a given bearer token and credential request.
        Supports API versioning via SWIYU-API-Version header. Returns the issued credential
        in JSON or JWT format.
      operationId: createCredential
      parameters:
      - name: Authorization
        in: header
        description: "Bearer token for authentication. Format: 'Bearer ..."
        required: true
        schema:
          type: string
      - name: DPoP
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialRequest"
          application/jwt:
            schema:
              type: string
              description: |
                An encoded JWT as described in RFC7519, with the claims as found in the unencrypted request
        required: true
      responses:
        "200":
          description: Credential issued successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialEndpointResponse"
        "202":
          description: Successful deferred credential. The credential will be issued
            later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeferredCredentialPendingResponse"
        "400":
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DpopError"
        "404":
          description: Referenced resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "405":
          description: HTTP method not allowed
          content:
            application/json:
              schema:
                type: string
            application/jwt:
              schema:
                type: string
        "409":
          description: The requested credential is currently locked by another operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "420":
          description: Non-standard status forwarded verbatim from the external Business
            Issuer renewal backend during a renewal request. Not a fixed contract
            value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Request validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: A renewal request for this credential is already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      security:
      - bearer-jwt: []
...
  /management/api/credentials:
    post:
      tags:
      - Credential API
      summary: Create a generic credential offer with the given content
      description: |
        Create a new credential offer, which can then be collected by the holder.
        The returned deep link has to be provided to the holder through another channel, for example as QR-Code.
        The credentialSubjectData can be a json object or a JWT, if the signer has been configured to perform data integrity checks.
        Returns both the ID used to interact with the offer and later issued VC, and the deep link to be provided to
      operationId: createCredential_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialOfferRequest"
        required: true
      responses:
        "200":
          description: Credential offer created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CredentialWithDeeplinkResponse"
        "400":
          description: |
            Bad request due to user content or internal call to external service like statuslist
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    ApiError:
      type: object
      description: Error response object
      properties:
        error:
          type: string
        error_description:
          type: string
        detail:
          type: string
        trace_id:
          type: string
      required:
      - error
...
    CredentialEndpointResponse:
      type: object
      properties:
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/CredentialObjectDto"
...
    DeferredCredentialPendingResponse:
      type: object
      properties:
        transaction_id:
          type: string
          description: String identifying the Deferred Issuance transaction for subsequent
            polling.
        interval:
          type: integer
          format: int64
          description: Minimum number of seconds the Wallet MUST wait before polling
            again.
      required:
      - interval
      - transaction_id
...
    NonceResponse:
      type: object
      properties:
        c_nonce:
          type: string
          description: String containing an unpredictable challenge to be used when
            creating a proof of possession of the key.
      required:
      - c_nonce
...
    OAuthToken:
      type: object
      properties:
        scope:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
          format: int64
      required:
      - access_token
    OauthAccessTokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          default: urn:ietf:params:oauth:grant-type:pre-authorized_code
          description: The type of grant being requested. Must be 'urn:ietf:params:oauth:grant-type:pre-authorized_code'
            or 'refresh_token.
          minLength: 1
        pre-authorized_code:
          type: string
        refresh_token:
          type: string
      required:
      - grant_type
...
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT